% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{alternativeEvents}
\alias{alternativeEvents}
\title{Filter Splicing Events by PSI Range}
\usage{
alternativeEvents(filteredList, minPsi, maxPsi)
}
\arguments{
\item{filteredList}{A list, typically the output from \code{\link{filterEvents}} or
\code{\link{getEvents}}.}

\item{minPsi}{A numeric value between 0 and 100, representing the minimum PSI threshold.}

\item{maxPsi}{A numeric value between 0 and 100, representing the maximum PSI threshold.}
}
\value{
A list containing:
       1. PSI filtered table based on the provided PSI range.
       2. Corresponding Qual filtered table.
       3. A table enumerating the number of events per type.
       4. A character vector, "Samples", derived from column names.
       This mirrors the structure of \code{\link{getDataset}},
       but with additional filtering based on the PSI range.
}
\description{
This function filters splicing events based on a specified PSI range.
Events with a PSI value outside of the provided range are removed.
}
\details{
The function filters out events whose PSI value (percentage spliced in) does not fall
within the [minPsi, maxPsi] range. Both bounds are inclusive.
}
\examples{
# Assuming you have a filteredList from filterEvents():
load_data <- getDataset(pathTables=NULL, tool="vast-tools")   # Replace with your actual data
data <- getEvents(load_data, tool="vast-tools")
filtered_data <- filterEvents(InputList = data, types = c("CE", "AA"), N = 10)
alternative_data <- alternativeEvents(filteredList = filtered_data, minPsi = 20, maxPsi = 80)

}
