% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVastTools.R
\name{alternativeVastTools}
\alias{alternativeVastTools}
\title{Filter PSI table (vast-tools) by alternativity
Filter previously quality-filtered PSI table from vast-tools to consider only events with PSIs between (and including) minPsi and maxPSI.}
\usage{
alternativeVastTools(filteredVTList, minPsi, maxPsi)
}
\arguments{
\item{filteredVTList}{List containing PSI and Qual (vast-tools) tables, obtained with filterVastTools()}

\item{minPsi}{(numeric) Minimum PSI to consider}

\item{maxPsi}{(numeric) Maximum PSI to consider}
}
\value{
List with: 1) filtered table PSI columns, 2) filtered table Qual columns, 3) table with number of events per type and 4) Samples (based on colnames)
}
\description{
Filter PSI table (vast-tools) by alternativity
Filter previously quality-filtered PSI table from vast-tools to consider only events with PSIs between (and including) minPsi and maxPSI.
}
\examples{
testTable <- betAS:::testTable
# filter table to consider only events all PSIs between 1 and 99
alternativeVastTools(testTable, minPsi = 1, maxPsi = 99)
}
