% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betas.R
\name{betas.lmr}
\alias{betas.lmr}
\title{Compute standardized beta coeffizients for robust linear regression models}
\usage{
betas.lmr(object, classic = FALSE)
}
\arguments{
\item{object}{A model.}

\item{classic}{classic cov.}
}
\value{
beta coeffizients.
}
\description{
Compute standardized beta coeffizients for robust linear regression models
}
\examples{
data <- pisa2012che
library(robust)

fit1 <- lmRob(MATH ~ ESCS + USEMATH, data)
betas.lmr(fit1)
}

