% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{Compute standardized beta coeffizients for linear regression models}
\usage{
betas(MOD)
}
\arguments{
\item{MOD}{A model.}
}
\value{
beta coeffizients.
}
\description{
Compute standardized beta coeffizients for linear regression models
}
\examples{
#load("pisa2012che.RData")
data <- pisa2012che

## linear regression models with numerical covariates only
fit1 <- lm(MATH ~ ESCS + USEMATH, data)
betas(fit1)

## linear regression models with numerical and factorial covariates
fit2 <- lm(MATH ~ ESCS + USEMATH + ST04Q01 + FAMSTRUC + ST28Q01, data)
betas(fit2)

## weighted linear regression models
fit3 <- lm(MATH ~ ESCS + USEMATH, data, weights = W_FSTUWT)
betas(fit3)

fit4 <- lm(MATH ~ ESCS + USEMATH + ST04Q01 + FAMSTRUC + ST28Q01, data, weights = W_FSTUWT)
betas(fit4)
}

