% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betASapp.R
\name{betASapp}
\alias{betASapp}
\title{Run betAS graphical interface}
\usage{
betASapp(...)
}
\arguments{
\item{...}{Arguments passed to shiny::runApp()}
}
\value{
betAS graphical interface (shiny app)
}
\description{
Run betAS graphical interface
}
