% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{CheckMinReads}
\alias{CheckMinReads}
\title{Verify Minimum Reads for a Given Event Across Samples}
\usage{
CheckMinReads(quals, N)
}
\arguments{
\item{quals}{A set of "Qual" columns (one per sample) for a given event.}

\item{N}{The minimum number of reads (inc + exc) for a given event in each sample.}
}
\value{
A boolean value indicating if all values of (inc + exc) across samples
        for a given event are greater than or equal to the specified minimum, \code{N}.
}
\description{
This function checks if the coverage associated to a given event (inc+exc)
across samples for a specific event meet a certain threshold.
}
\examples{
# Assuming you have a set of "Qual" columns for a given event:
quals_data <- c(...) # Replace with your actual data
min_reads_check <- CheckMinReads(quals = quals_data, N = 5)
if (min_reads_check) {
  print("All samples meet the minimum reads criteria.")
} else {
  print("Not all samples meet the minimum reads criteria.")
}

}
