% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{convertCols}
\alias{convertCols}
\title{Convert between column names and indexes}
\usage{
convertCols(psitable, index)
}
\arguments{
\item{psitable}{PSI table}

\item{index}{(vector) column names or indexes}
}
\value{
named vector with column indexes and corresponding sample names (colnames)
}
\description{
Convert between column names and indexes
}
\examples{
testTable <- betAS:::testTable
psiTable   <- testTable$PSI
# get named vector with column indexes and corresponding sample names
convertCols(psiTable, index = c(10,11,13))
convertCols(psiTable, index = c("ERR2598269", "ERR2598270", "ERR2598271"))
}
