% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointBetas_nofitting.R
\name{estimateFDR}
\alias{estimateFDR}
\title{Estimate FDR associated with the obtained delta PSI (vast-tools)
Takes samples from two groups, generates two indBetas objects and estimates the FDR of the obtained delta PSI by simulating nsim delta PSI under the null hypothesis that all samples have the same original PSI.}
\usage{
estimateFDR(
  indBetasA,
  indBetasB,
  groupsAB,
  nsim,
  seed = TRUE,
  CoverageWeight = FALSE
)
}
\arguments{
\item{indBetasA}{(indBetas) individual betAS object of group A samples}

\item{indBetasB}{(indBetas) individual betAS object of group B samples}

\item{groupsAB}{(vector) vector providing labels for groups A and B (in this order)}

\item{nsim}{(numeric) number of simulations}

\item{seed}{(boolean) Boolean indicating if seed should be fixed. Default is TRUE}

\item{CoverageWeight}{(boolean)  Boolean indicating if number of points emitted in beta distributions should be weighted by the coverage of each sample. Default is FALSE}
}
\value{
jntBetas object list with: 1) vector of simulated delta PSIs; 2) original median PSI found, 3) FDR and 4) plot with the obtained FDR compared to the simulated
}
\description{
Estimate FDR associated with the obtained delta PSI (vast-tools)
Takes samples from two groups, generates two indBetas objects and estimates the FDR of the obtained delta PSI by simulating nsim delta PSI under the null hypothesis that all samples have the same original PSI.
}
