% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{filterEvents}
\alias{filterEvents}
\title{Filter Splicing Events by Quality, Event Type, and Minimum Reads}
\usage{
filterEvents(InputList, types = NULL, N = 10)
}
\arguments{
\item{InputList}{A list, typically the output from \code{\link{getEvents}}.}

\item{types}{A character vector of column letter codes for alternative event types.
For Whippet, valid types include c("CE", "AA", "AD", "IR", "TS", "TE", "AF", "AL", "BS")
as described in \url{https://github.com/timbitz/Whippet.jl}.
For vast-tools, valid types include c("Alt3", "Alt5", "ANN", "C1", "C2", "C3", "IR-C", "IR-S", "MIC", "S")
as described in \url{https://github.com/vastgroup/vast-tools}.
For rMATS, as each table only contains one event, this parameter should be set to NULL.}

\item{N}{An integer, representing the minimum number of reads (inc + exc) for a given event in each sample.}
}
\value{
A list containing:
       1. Psi filtered table.
       2. Qual filtered table.
       3. A table enumerating the number of events per type.
       4. A character vector, "Samples", derived from column names.
       Essentially, it mirrors the structure of \code{\link{getDataset}},
       but with events filtered by coverage and type.
}
\description{
This function filters splicing events based on quality (presence of NAs),
specified event types, and a minimum read count threshold.
}
\examples{
# Assuming you have an InputList from getEvents():
input_data <- getEvents(...) # Replace with your actual data
# Filtering whippet events, to isolate only CE and AA events with a minimum of 10 reads in each sample
filtered_data <- filterEvents(InputList = input_data, types = c("CE", "AA"), N = 10)

}
