% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointBetas_nofitting.R
\name{fStatistic_2Groups_PerEvent}
\alias{fStatistic_2Groups_PerEvent}
\title{Joint betAS (vast-tools) - F-statistic
Takes samples from two groups, generates two indBetas objects and estimates the F-statistic using the ration between differences between beta distributions of groups A and B and within each group. To be applied per row of Qual table,
using pblapply() or similar.}
\usage{
fStatistic_2Groups_PerEvent(
  eventPos,
  psitable,
  qualtable,
  npoints,
  colsA,
  colsB,
  labA,
  labB,
  maxDevTable,
  seed = TRUE,
  CoverageWeight = FALSE
)
}
\arguments{
\item{seed}{(boolean) Boolean indicating if seed should be fixed. Default is TRUE}

\item{CoverageWeight}{(boolean)  Boolean indicating if number of points emitted in beta distributions should be weighted by the coverage of each sample. Default is FALSE}
}
\description{
Joint betAS (vast-tools) - F-statistic
Takes samples from two groups, generates two indBetas objects and estimates the F-statistic using the ration between differences between beta distributions of groups A and B and within each group. To be applied per row of Qual table,
using pblapply() or similar.
}
