% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointBetas_nofitting.R
\name{generalisedGroupsBetas}
\alias{generalisedGroupsBetas}
\title{Estimate multiple-groups differential alternative splicing statistics
Takes group-wise beta distributions and estimates the emitted values differences between and within groups, from which statistics of differential alternative splicing across groups (Pdiff and F-statistic).}
\usage{
generalisedGroupsBetas(
  eventPos,
  indList,
  samplesList,
  groupNames,
  seed = TRUE,
  CoverageWeight = FALSE
)
}
\arguments{
\item{eventPos}{(numeric) position of event under analysis in provided list}

\item{indList}{(list) list of individual betAS objects per group}

\item{samplesList}{(list) list of vectors of samples corresponding to each group}

\item{groupNames}{(character vector) vector of group names in indList}
}
\value{
dirGroupBetas object list with: 1) samples considered; 2) group per sample, 3) F-stat, 4) P(|betweens| > |withins|), 5) Transformed Pdiff, 6) Median of between distances and 7) Difference between median(|betweens|) to median(|withins|)
}
\description{
Estimate multiple-groups differential alternative splicing statistics
Takes group-wise beta distributions and estimates the emitted values differences between and within groups, from which statistics of differential alternative splicing across groups (Pdiff and F-statistic).
}
