% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{getDataset}
\alias{getDataset}
\title{Retrieve Inclusion Tables for Specific Tools}
\usage{
getDataset(pathTables = NULL, tool)
}
\arguments{
\item{pathTables}{Character or list. Specifies the source of the dataset:
\itemize{
  \item{\strong{vast-tools:}}{Should be a path pointing to a \code{*INCLUSION_LEVELS_FULL*.tab} file.}
  \item{\strong{rMATS:}}{Should be a path pointing to a \code{*MATS.JC.txt} file.}
  \item{\strong{whippet:}}{Should be a list of paths (minimum of 2) pointing to \code{*.psi.gz} files.}
}
If \code{NULL}, the function will load the default betAS test datasets for the chosen tool.}

\item{tool}{A character string specifying the desired tool for which the dataset
should be fetched. Acceptable values are "vast-tools", "whippet", or "rMATS".}
}
\value{
Returns a data frame if the tool is "rMATS" or "vast-tools". If the tool is
       "whippet", the function returns a list of data frames.
}
\description{
This function retrieves the inclusion tables specific to the chosen tool for further analyses.
}
\examples{
# Example using the vast-tools tool with a single path
# getDataset(pathTables = "/path/to/vast_tools_file.txt", tool = "vast-tools")

# Example using the whippet tool with multiple paths
# getDataset(pathTables = list("/path/to/whippet_file1.txt", "/path/to/whippet_file2.txt"), tool = "whippet")

# Loading the betAS test datasets for rMATS
getDataset(pathTables = NULL, tool = "rMATS")

}
