% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{getEvents}
\alias{getEvents}
\title{Format Inclusion Tables into a Standard betAS Format}
\usage{
getEvents(incTable, tool)
}
\arguments{
\item{incTable}{A data frame equivalent object (or list of data frames, when using tool=\code{whippet},
representing the inclusion table for the chosen tool. This is typically the output of the `getDataset`
function when not using the `betASapp()` shiny app.}

\item{tool}{A character string specifying the tool from which the incTable is derived.
It can take one of three values: "vast-tools", "whippet", or "rMATS".}
}
\value{
A list containing four data frames:
  \itemize{
    \item \strong{PSI}: A data frame with information regarding the splicing
          events and their inclusion levels.
    \item \strong{Qual}: A data frame detailing the events and information
          related to the coverage associated with each event. This representation
          follows the vast-tools representation of inclusion (inc) and
          exclusion (exc) reads.
    \item \strong{EventsPerType}: A named table summarising the number of
          events per type.
    \item \strong{Samples}: Vector containing the names of the samples considered.
  }
}
\description{
This function formats inclusion tables from one of the three specified tools into
a standard format that is compatible with betAS.
}
\examples{
# Assuming you have already loaded or created an incTable from rMATS using the getDataset() function:
rMATS_incTable <- getDataset(pathTables=NULL, tool="rMATS") # substitute pathTables with the path to your data
formattedTable_rMATS <- getEvents(incTable = rMATS_incTable, tool = "rMATS")

# Similarly, for vast-tools:
vasttools_incTable <- getDataset(pathTables=NULL, tool="vast-tools")
formattedTable_vasttools <- getEvents(incTable = vasttools_incTable, tool = "vast-tools")

# And for whippet:
whippet_incTable <- getDataset(pathTables=NULL, tool="whippet")
formattedTable_whippet <- getEvents(incTable = whippet_incTable, tool = "whippet")

}
