% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{getrMATS}
\alias{getrMATS}
\title{Format rMATS Inclusion Tables for betAS Compatibility}
\usage{
getrMATS(incTable)
}
\arguments{
\item{incTable}{A data frame or similar object representing
the inclusion table specifically from rMATS. This is
typically the output of the `getDataset` function when not
using the `betASapp()` shiny app. Can also be a data frame obtained
by reading a \code{*MATS.JC.txt} file}
}
\value{
A list containing:
       \itemize{
       \item{"PSI"}{A data frame containing information regarding the splicing events
                     and their inclusion level.}
       \item{"Qual"}{A data frame with information on the events as well as information
                      regarding the coverage associated with each, emulating rMATS
                      way of representing inc and exc reads.}
       \item{"EventsPerType"}{A named table summarizing the number of events per type.}
       \item{"Samples"}{A character vector representing the names of the samples considered.}
       }
}
\description{
This function reformats inclusion tables specifically from rMATS
into a standardized format that is compatible with betAS.
}
\details{
rMATS is one of several splicing tools that output inclusion tables.
For the sake of consistent analysis within betAS, these tables need to be
standardized. This function processes rMATS output and structures it
in a way that betAS can work with, ensuring compatibility and ease of further analysis.
}
