% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{getWhippet}
\alias{getWhippet}
\title{Format Whippet Inclusion Tables for betAS Compatibility}
\usage{
getWhippet(listIncTables)
}
\arguments{
\item{incTable}{A list of paths (minimum length of 2) pointing to
the inclusion tables specifically from Whippet. This is
typically the output of the `getDataset` function when not
using the `betASapp()` shiny app. Can also be a list of data frames
obtained by reading a list of \code{*.psi.gz} files.}
}
\value{
A list containing:
       \itemize{
       \item{"PSI"}{A data frame containing information regarding the splicing events
                     and their inclusion level.}
       \item{"Qual"}{A data frame with information on the events as well as information
                      regarding the coverage associated with each, emulating Whippet
                      way of representing inc and exc reads.}
       \item{"EventsPerType"}{A named table summarizing the number of events per type.}
       \item{"Samples"}{A character vector representing the names of the samples considered.}
       }
}
\description{
This function reformats inclusion tables specifically from Whippet
into a standardized format that is compatible with betAS.
}
\details{
Whippet is one of several splicing tools that output inclusion tables.
For the sake of consistent analysis within betAS, these tables need to be
standardized. This function processes Whippet output and structures it
in a way that betAS can work with, ensuring compatibility and ease of further analysis.
}
