% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hc_theme_smpl_tailored}
\alias{hc_theme_smpl_tailored}
\title{Adapted from highcharts themes: "simple theme for highcharts"}
\usage{
hc_theme_smpl_tailored(...)
}
\arguments{
\item{...}{}
}
\description{
Adapted from highcharts themes: "simple theme for highcharts"
}
