% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualBetas_nofitting_incr.R
\name{individualBetas_nofitting_incr}
\alias{individualBetas_nofitting_incr}
\title{Individual betAS
Generates indpoints from a beta distribution for a group of samples, with shape parameters based on inc and exc from Qual columns (vast-tools) table; requires a table with the first 6 columns and the “.Q” columns from vast-tools’ output table “INCLUSION”, the columns (cols) to consider; requires a maximum deviation reference table maxdevRefTable to be used to increment inc or exc if any  is equal to zero.}
\usage{
individualBetas_nofitting_incr(
  table,
  cols,
  indpoints,
  maxdevRefTable,
  seed = TRUE,
  CoverageWeight = FALSE
)
}
\arguments{
\item{table}{(data.frame) row from vast-tools output INCLUSION table with the 6 identifier columns and “.Q” columns for each sample. Function is applied per row of INCLUSION table using pblapply() or similar.}

\item{cols}{(integer) vector with columns of table to consider}

\item{indpoints}{(integer) number of points/observations emitted from the beta distribution}

\item{maxdevRefTable}{(data.frame) reference data frame with maximum increment values per coverage (inc+exc) to avoid emitted values to be artifitially beyond a certain threshold}

\item{seed}{(boolean) Boolean indicating if seed should be fixed. Default is TRUE}

\item{CoverageWeight}{(boolean)  Boolean indicating if number of points emitted in beta distributions should be weighted by the coverage of each sample. Default is FALSE}
}
\value{
indBetas object list with, for event: 1) inc = number of inclusion reads, 2) exc = number of exclusion reads, 3) PSIs, 4 and 5) corrected inc and exc (using a small increment to avoid zeros), 6) corrected PSIs  (based on corrected inc and exc), 7) points emitted from beta distribution, 8) mean and 9) median of the distribution.
}
\description{
Individual betAS
Generates indpoints from a beta distribution for a group of samples, with shape parameters based on inc and exc from Qual columns (vast-tools) table; requires a table with the first 6 columns and the “.Q” columns from vast-tools’ output table “INCLUSION”, the columns (cols) to consider; requires a maximum deviation reference table maxdevRefTable to be used to increment inc or exc if any  is equal to zero.
}
