% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointBetas_nofitting.R
\name{jointBetas_nofitting_Fast0}
\alias{jointBetas_nofitting_Fast0}
\title{Joint betAS (vast-tools) - fast version
Takes two indBetas objects and performs differential alternative splicing analysis considering the beta distributions between groups A and B. To be applied per row of Qual table,
i.e. per element of lists with indBetas event-wise results (list names = events IDs) using pblapply() or similar.}
\usage{
jointBetas_nofitting_Fast0(indBetasA, indBetasB, groupsAB)
}
\arguments{
\item{indBetasA}{(indBetas) individual betAS object of group A samples}

\item{indBetasB}{(indBetas) individual betAS object of group B samples}

\item{groupsAB}{(vector) vector providing labels for groups A and B (in this order)}
}
\value{
jntBetas object list with: 1) P(x = 0) and 2) Flipped P(x = 0), i.e. Pdiff
}
\description{
Joint betAS (vast-tools) - fast version
Takes two indBetas objects and performs differential alternative splicing analysis considering the beta distributions between groups A and B. To be applied per row of Qual table,
i.e. per element of lists with indBetas event-wise results (list names = events IDs) using pblapply() or similar.
}
