% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betas.R
\name{my.lmr.beta}
\alias{my.lmr.beta}
\title{Compute standardized beta coeffizients for robust linear regression models}
\usage{
my.lmr.beta(object, classic = FALSE)
}
\arguments{
\item{object}{A model.}

\item{classic}{classic cov.}
}
\value{
beta coeffizients.
}
\description{
Compute standardized beta coeffizients for robust linear regression models
}

