% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa2012che}
\alias{pisa2012che}
\title{Subset of the PISA 2012 students data set.}
\format{A data frame with 100 rows and 8 variables:
\describe{
  \item{MATH}{Mean score in mathematics}
  \item{ESCS}{Index of economic, social and cultural status}
  \item{ST04Q01}{Gender}
  \item{FAMSTRUC}{Family Structure}
  \item{ST28Q01}{How many books at home}
  \item{IC08Q01}{One player games}
  \item{USEMATH}{Use of ICT in Mathematic Lessons}
  \item{W_FSTUWT}{FINAL STUDENT WEIGHT}
}}
\source{
\url{http://pisa2012.acer.edu.au/downloads.php}
}
\usage{
pisa2012che
}
\description{
A data set containing Math scores and some covariates and weights
for students from Switzerland.
}
\keyword{datasets}

