% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDensitiesFromEventObjList}
\alias{plotDensitiesFromEventObjList}
\title{Plot individual beta distributions per sample}
\usage{
plotDensitiesFromEventObjList(eventObjList, colorA, colorB)
}
\arguments{
\item{eventObjList}{}
}
\description{
Plot individual beta distributions per sample
}
