% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFDRFromEventObjList}
\alias{plotFDRFromEventObjList}
\title{Plot FDR explanation plot per sample}
\usage{
plotFDRFromEventObjList(eventObjList)
}
\arguments{
\item{eventObjList}{}
}
\description{
Plot FDR explanation plot per sample
}
