% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFstatFromEventObjList}
\alias{plotFstatFromEventObjList}
\title{Plot Fstat explanation plot per sample}
\usage{
plotFstatFromEventObjList(eventObjList)
}
\arguments{
\item{eventObjList}{}
}
\description{
Plot Fstat explanation plot per sample
}
