% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFstatFromEventObjListMultiple}
\alias{plotFstatFromEventObjListMultiple}
\title{Plot Fstat explanation plot per sample}
\usage{
plotFstatFromEventObjListMultiple(eventObjList)
}
\arguments{
\item{eventObjList}{}
}
\description{
Plot Fstat explanation plot per sample
}
