% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotIndividualDensities}
\alias{plotIndividualDensities}
\title{Plot individual beta distributions per sample}
\usage{
plotIndividualDensities(
  eventID,
  npoints,
  psitable,
  qualtable,
  colsA,
  colsB,
  labA,
  labB,
  colorA,
  colorB,
  maxDevTable,
  seed = TRUE,
  CoverageWeight = FALSE
)
}
\arguments{
\item{seed}{(boolean) Boolean indicating if seed should be fixed. Default is TRUE}

\item{CoverageWeight}{(boolean)  Boolean indicating if number of points emitted in beta distributions should be weighted by the coverage of each sample. Default is FALSE}
}
\description{
Plot individual beta distributions per sample
}
