% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotIndividualDensitiesList}
\alias{plotIndividualDensitiesList}
\title{Plot individual beta distributions per sample from groups list}
\usage{
plotIndividualDensitiesList(
  eventID,
  npoints,
  psitable,
  qualtable,
  groupList,
  maxDevTable,
  seed = TRUE,
  CoverageWeight = FALSE
)
}
\arguments{
\item{eventID}{(character) event ID}

\item{npoints}{number of points emitted by each beta distribution}

\item{psitable}{PSI table}

\item{qualtable}{Qual table}

\item{groupList}{group list}

\item{maxDevTable}{table with increments to be summed to zero-valued reads, per coverage}

\item{seed}{(boolean) Boolean indicating if seed should be fixed. Default is TRUE}

\item{CoverageWeight}{(boolean)  Boolean indicating if number of points emitted in beta distributions should be weighted by the coverage of each sample. Default is FALSE}
}
\value{
ggplot density plot grid
}
\description{
Plot individual beta distributions per sample from groups list
}
\examples{
testTable <- betAS:::testTable
maxDevSimulationN100  <- betAS:::maxDevSimulationN100
psiTable   <- testTable$PSI
qualTable  <- testTable$Qual
testGroups <- list()
testGroups[[length(testGroups)+1]] <- list(name = "GroupA", samples = c("ERR2598266", "ERR2598267", "ERR2598268"), color = "#FF9AA2")
testGroups[[length(testGroups)+1]] <- list(name = "GroupB", samples = c("ERR2598270", "ERR2598307", "ERR2598351"), color = "#FFB7B2")
plotIndividualDensitiesList(eventID = "HsaEX0019479", npoints = 500, psitable = psiTable, qualtable = qualTable, groupList = testGroups, maxDevTable = maxDevSimulationN100)
}
