% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPDiffFromEventObjList}
\alias{plotPDiffFromEventObjList}
\title{Plot Pdiff explanation plot per sample}
\usage{
plotPDiffFromEventObjList(eventObjList)
}
\arguments{
\item{eventObjList}{}
}
\description{
Plot Pdiff explanation plot per sample
}
