% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot betAS volcano plot}
\usage{
plotVolcano(betasTable, labA, labB, basalColor, interestColor)
}
\arguments{
\item{betasTable}{data table generated with prepareTableVolcano}

\item{labA}{group A label}

\item{labB}{group B label}

\item{basalColor}{general color for points (events)}

\item{interestColor}{color for highlighted points (events)}
}
\value{
ggplot scatterplot
}
\description{
Plot betAS volcano plot
}
\examples{
testTable <- betAS:::testTable
maxDevSimulationN100  <- betAS:::maxDevSimulationN100
psiTable   <- testTable$PSI
qualTable  <- testTable$Qual
testGroups <- list()
testGroups[[length(testGroups)+1]] <- list(name = "GroupA", samples = c("ERR2598266", "ERR2598267", "ERR2598268"), color = "#FF9AA2")
testGroups[[length(testGroups)+1]] <- list(name = "GroupB", samples = c("ERR2598270", "ERR2598307", "ERR2598351"), color = "#FFB7B2")
groupA    <- "GroupA"
groupB    <- "GroupB"
samplesA   <- testGroups[[1]]$samples
samplesB   <- testGroups[[2]]$samples
colsGroupA    <- convertCols(psiTable, samplesA)
colsGroupB    <- convertCols(psiTable, samplesB)
volcanoTable <- prepareTableVolcano(psitable = psiTable, qualtable = qualTable, npoints = 500, colsA = colsGroupA, colsB = colsGroupB, labA = groupA, labB = groupB, basalColor = "#89C0AE", interestColor = "#E69A9C", maxDevTable = maxDevSimulationN100)
plotVolcano(betasTable = volcanoTable, labA = groupA, labB = groupB, basalColor = "#89C0AE", interestColor = "#E69A9C")
}
