% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVolcano_MultipleGroups_Fstat}
\alias{plotVolcano_MultipleGroups_Fstat}
\title{Plot betAS volcano plot (with Fstat as y-axis metric)}
\usage{
plotVolcano_MultipleGroups_Fstat(betasTable)
}
\arguments{
\item{betasTable}{data table generated with prepareTableVolcanoMultipleGroups}

\item{basalColor}{general color for points (events)}

\item{interestColor}{color for highlighted points (events)}
}
\value{
ggplot scatterplot
}
\description{
Plot betAS volcano plot (with Fstat as y-axis metric)
}
