% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVolcanoFDR}
\alias{plotVolcanoFDR}
\title{Plot betAS volcano plot with the estimated FPR}
\usage{
plotVolcanoFDR(betasTable, labA, labB, basalColor, interestColor)
}
\arguments{
\item{interestColor}{}
}
\description{
Plot betAS volcano plot with the estimated FPR
}
