% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVolcanoFstat}
\alias{plotVolcanoFstat}
\title{Plot betAS volcano plot with F-statistic for the y-axis}
\usage{
plotVolcanoFstat(betasTable, labA, labB, basalColor, interestColor)
}
\arguments{
\item{interestColor}{}
}
\description{
Plot betAS volcano plot with F-statistic for the y-axis
}
