% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{preparePieForVastToolsCOMPLEX}
\alias{preparePieForVastToolsCOMPLEX}
\title{Prepare pie chart}
\usage{
preparePieForVastToolsCOMPLEX(table)
}
\arguments{
\item{table}{PSI table}
}
\value{
highcharts pie chart
}
\description{
Prepare pie chart
}
\examples{
testTable <- betAS:::testTable
preparePieForVastToolsCOMPLEX(table = testTable$PSI)
}
