% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepareTableEvent}
\alias{prepareTableEvent}
\title{Prepare and return table used for betAS event-wise plots}
\usage{
prepareTableEvent(
  eventID,
  psitable,
  qualtable,
  npoints,
  colsA,
  colsB,
  labA,
  labB,
  basalColor,
  interestColor,
  maxDevTable,
  nsim
)
}
\arguments{
\item{nsim}{}
}
\description{
Prepare and return table used for betAS event-wise plots
}
