% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepareTableEventMultiple}
\alias{prepareTableEventMultiple}
\title{Prepare and return table used for betAS event-wise plots (multiple-group section)}
\usage{
prepareTableEventMultiple(
  eventID,
  psitable,
  qualtable,
  groupList,
  npoints,
  maxDevTable
)
}
\arguments{
\item{maxDevTable}{}
}
\description{
Prepare and return table used for betAS event-wise plots (multiple-group section)
}
