% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepareTableVolcanoFDR}
\alias{prepareTableVolcanoFDR}
\title{Prepare and return table used for betAS volcano plot with the estimated FPR}
\usage{
prepareTableVolcanoFDR(
  psitable,
  qualtable,
  npoints,
  colsA,
  colsB,
  labA,
  labB,
  basalColor,
  interestColor,
  maxDevTable,
  nsim,
  seed = TRUE,
  CoverageWeight = FALSE
)
}
\arguments{
\item{seed}{(boolean) Boolean indicating if seed should be fixed. Default is TRUE}

\item{CoverageWeight}{(boolean)  Boolean indicating if number of points emitted in beta distributions should be weighted by the coverage of each sample. Default is FALSE}
}
\description{
Prepare and return table used for betAS volcano plot with the estimated FPR
}
