% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepareTableVolcanoFstat}
\alias{prepareTableVolcanoFstat}
\title{Prepare and return table used for betAS volcano plot with F-statistic for the y-axis}
\usage{
prepareTableVolcanoFstat(
  psitable,
  qualtable,
  npoints,
  colsA,
  colsB,
  labA,
  labB,
  basalColor,
  interestColor,
  maxDevTable
)
}
\arguments{
\item{interestColor}{}
}
\description{
Prepare and return table used for betAS volcano plot with F-statistic for the y-axis
}
