% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepareTableVolcanoMultipleGroups}
\alias{prepareTableVolcanoMultipleGroups}
\title{Prepare and return table used for betAS volcano plot (multiple group analysis)}
\usage{
prepareTableVolcanoMultipleGroups(
  psitable,
  qualtable,
  groupList,
  npoints,
  maxDevTable
)
}
\arguments{
\item{maxDevTable}{}
}
\description{
Prepare and return table used for betAS volcano plot (multiple group analysis)
}
