% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rMATS_data_mouse}
\alias{rMATS_data_mouse}
\title{rMATS example dataset}
\format{
A data frame
}
\source{
https://www.ncbi.nlm.nih.gov/bioproject/PRJNA185305/
}
\usage{
rMATS_data_mouse
}
\description{
This dataset is a subset of the
[Deep transcriptional profiling of longitudinal changes during neurogenesis and network maturation in vivo](https://www.ncbi.nlm.nih.gov/bioproject/PRJNA185305/)
public mouse dataset.
}
\keyword{datasets}
