% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rMATS_metadata_mouse}
\alias{rMATS_metadata_mouse}
\title{rMATS example metadata}
\format{
A data frame (metadata)
}
\source{
https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6814/
}
\usage{
rMATS_metadata_mouse
}
\description{
This dataset is the metadata of a subset of the
[Deep transcriptional profiling of longitudinal changes during neurogenesis and network maturation in vivo](https://www.ncbi.nlm.nih.gov/bioproject/PRJNA185305/)
public mouse dataset.
}
\keyword{datasets}
