% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VT1_metadata_human}
\alias{VT1_metadata_human}
\title{vast-tools 1 example metadata}
\format{
A data frame
}
\source{
https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6814/
}
\usage{
VT1_metadata_human
}
\description{
This dataset is the metadata of the
[Human RNA-seq time-series of the development of seven major organs](" https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6814/")
public human dataset.
}
\keyword{datasets}
