% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{CheckMinReads}
\alias{CheckMinReads}
\title{Verify Minimum Reads for a Given Event Across Samples}
\usage{
CheckMinReads(quals, N)
}
\arguments{
\item{quals}{A set of "Qual" columns (one per sample) for a given event.}

\item{N}{The minimum number of reads (inc + exc) for a given event in each sample.}
}
\value{
A boolean value indicating if all values of (inc + exc) across samples
        for a given event are greater than or equal to the specified minimum, \code{N}.
}
\description{
This function checks if the coverage associated to a given event (inc+exc)
across samples for a specific event meet a certain threshold.
}
