% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotIndividualViolinsList}
\alias{plotIndividualViolinsList}
\title{Plot individual beta distributions as violin plots per sample from groups list}
\usage{
plotIndividualViolinsList(
  eventID,
  npoints,
  psitable,
  qualtable,
  groupList,
  maxDevTable
)
}
\arguments{
\item{eventID}{(character) event ID}

\item{npoints}{number of points emitted by each beta distribution}

\item{psitable}{PSI table}

\item{qualtable}{Qual table}

\item{groupList}{group list}

\item{maxDevTable}{table with increments to be summed to zero-valued reads, per coverage}
}
\value{
ggplot density plot grid
}
\description{
Plot individual beta distributions as violin plots per sample from groups list
}
