% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepareTableVolcanoFDR}
\alias{prepareTableVolcanoFDR}
\title{Prepare and return table used for betAS volcano plot with the estimated FDR}
\usage{
prepareTableVolcanoFDR(
  psitable,
  qualtable,
  npoints,
  colsA,
  colsB,
  labA,
  labB,
  basalColor,
  interestColor,
  maxDevTable,
  nsim
)
}
\arguments{
\item{interestColor}{}
}
\description{
Prepare and return table used for betAS volcano plot with the estimated FDR
}
